pyproject.toml
[project]
name = "lucid-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "pyjwt>=2.10.1",
    "redis>=5.2.1",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.0",
]
